Created file: gui_components.py
Moved classes for gui components from gui_main.py to gui_components.py
